dlna: split into player + provider.py
